+++ src/main/scala/lisa/mathematics/Ordinals.scala
@@ -2,7 +2,6 @@ package lisa.mathematics
 
 import lisa.automation.kernel.OLPropositionalSolver.Tautology
 import lisa.automation.kernel.SimplePropositionalSolver.*
-import lisa.automation.kernel.SimpleSimplifier.*
 import lisa.automation.settheory.SetTheoryTactics.*
 import lisa.kernel.proof.SequentCalculus as SC
 import lisa.mathematics.settheory.SetTheory.*
